Adds 14 new planetary cultures and 1 religion
Adds cultures for the following planets: Tersten, Lordania, Lorriman, Yuklid, Cinu.
Splits Terran into two cultures - Traditional and Integrated. Traditional are your Terran Terrans, and Integrated are those from planets occupied by Terra and thus more likely to be refugees and seen on the Torch.  This is to reflect the difference between loyal Terrans and disloyal Terrans, whom we see in-game.
Changes the wording on Spiritualism to reflect Yuklid Spiritualism.