Include missing rank for supply department
Before this, the supply department was missing a role for the E-5 rank for fleet. The Deck Technician could be E-2 to E-4, and the Deck Chief could be E-6 to E-8. I thought this was rather weird and probably unrealistic, because what would a hypothetical E-4 who is promoted to E-5 then do? Change departments for one rank? Not sure. So I just added E-5 to deck technician.