Caps medtechs at Trained Anatomy
Playing whackamole with cost adjustments gets tiring.
There's no legit uses corpsmen need more than Trained anatomy for (or even Basic tbh).
Only place it's used in is CPR, for speed of it, and for effectiveness highest of Anatomy and Medicine is used, so they can just take high Medicine skill if they want to be turbopumpers.