Engine Room Emergency Vents
:cl: Boznar
maptweak: Adds large vents to the sides of the engine room. These can be opened with a button in the emergency box in the control room.
maptweak: Renamed all of the glass box emergency buttons and added descriptions.
maptweak: Engine hatch bolt control button has been moved to the inside of the engine room next to the hatches.
/:cl: