Fixes invisible vines
It happens when something changes plant icon without calling this proc to udpate growth stages.
I can just plug couple places that I found to do that, but I think this is way more robust solutino not relying on people knowing they need to do additional call for THIS trait.