Gas overlay tweaks
Overlay alpha now changes based on concentration and pressure.
Changed icon to one with less pronounced patterns, to be easier on eyes when it's pased across entire screen.

Codeside moved overlays from global gas data to per-mixture (effectively per-zone because proc that generates them is only called by zone). They are cached and only generated when needed, so shouldn't be that bad.