Axes tcomm relays
Only mapped one was on centcomm, all the presets were unused. Mostly just makes it not buildable anymore.

The reason is that it's legitimate reason to exist is gone - it was used to make comms work on all decks by having a relay on each one. Nowadays tcomms machines handle connected zlevels naturally and have no need for this machien (cue it being unmapped).

Only use it still has is circumventing no-comms-to-away-mission thing if you build it on a shuttle. Do not want.

Gives ERT long range holopads in briefing room and their shuttle.