Ascent Voidsuits, Mining, and Chemistry
:cl: Boznar
maptweak: Adds mineral processing, chemistry, and cages to the ascent seedship.
rscadd: Adds voidsuits for Alates. Icons and code credit to Zuhayr.
/:cl:

Revives the voidsuits that were going to be added in #26947. Icons and modules done by MistakeNot, and are basically copy-paste. The mapping I took over and did from scratch. The voidsuits are mapped in addition to the hardsuits.