Tweak to bullet special damaging
Bullets had extra chance to embed or sever a vessel. Issue with that was, it was
 outside of the damage handling, and probabilities generally ignored armor, so you ALWAYS had severed artery, regardless of what armor and what bullet you were shot with.

Now it's moved to a new generic proc of obj, that organ calls if it was damaged by that object and wound was created.
Now it has to actually damage things, the damage is checked, no embeds in bruises, and armor can actually stop them now if you're lucky.
It also won't bloody things if it didn't actually cause a bloody wound