Fixes unathi suffocating in populated rooms
Recent changes made to unathi breathing pressure resulted in them suffocating in rooms with anything less than a perfect ratio of oxygen (21% being default), in rooms with 2 or more people this will drop to 20.9% and start giving the unathi in the room suffocation warnings, and at lower amounts they will start suffocating. 

This fixes that.