New Ship: UNSC Light Brigade.
Adds a new UNSC ship, the UNSC Light Brigade, this is a one Z level shipmap loosely based off of a Halberd Class Destroyer.

Also fixes several minor things on the ODP, tweaks the description and fixes one includes file.

Vessel was tested replacing the Iron Will and worked without issue. As is currently shown however, the new ship is NOT in use, and is just a "ready to go" map in the map folder.