New Covenant Ship.
Remodels the entire interior of the covenant ship. Name is pending, but this new ship and all its needed files can be found in "SDV To Be Named".

Also updates the UNSC Light Brigade to have hull segments more in line with the new covenant vessel.

Can be reviewed already, once a name is chosen by the covenant, I can alter the area and overmap names.