Drink expansion -- Adds 28 new drinks & recipes
Fixes two drink recipes, coffee changes now only produce 3u instead of 4u

Aloe don't work as intended, this fixes it.

Decreases drink output of all coffees/teas to a base 3 instead of 4, as 30 is the maximum a mug can hold, so it can be mixed in the lounge without error and fill a cup.  Changes https://github.com/Baystation12/Baystation12/pull/27410 and approved by Sierra.