Xenobotany Atmospheric Chamber Fix
:cl: Lindhrive
maptweak: The atmospheric chamber in Xenobotany works now. The console outside can be used to pump gasses in or out.
/:cl:

The atmospheric chamber in Xenobotany was not working as intended. The vent has been replaced with an injector and the scrubber with a tank-vent, and the console has been configured to speak with them. It now works like the chamber in the Miscellaneous lab, only with the vent defaulted to off so as not to pop any unsuspecting scientist lungs.