Removes ability to construct tele-circuits
:cl:
rscdel: Removes the ability to construct tele-circuits
/:cl:
Due to a week-long (possibly longer) abuse of portable teleporters (despite repeated and unreserved in-character punishment), I am removing the ability to create them in the integrated circuits menu. The Captain and ERT hand teleporters will be unaffected by this change, and the component will still exist in the game, just without any way of constructing them.

If there are any questions or concerns about my choice and decision to remove this feature, you are welcome to reach out to me on discord at @Ozwell#3029, where I will happily discuss this.

* Removes the ability to make the component necessary for a teleporter circuit in the integrated circuit menu.