Give AI Master IT
:cl:
tweak: AI now has master information technology skill
/:cl:

This PR does two things.
1. Add a system to override the AI's default (expert) skill levels.
2. Give the AI master Information Technology.

Every time I have to inform Engineering that the AI cannot run `locate`
they are suprised and skeptical. The only option is to wait for a CE to
join, and hope the CE happens to have master IT, which oftentimes they
don't.

If there is a reason that the AI should not have `locate` please let me
know, and I will revise this PR to only add the skill adjustment
feature, but do nothing with it.