Salvage Gantry Human Provocateurs
Initial commit of a Human Provocateur role, based on a Shuttle Salvage Gantry with attached Habitation module
The hull of a Yacht was attached center-line to provide thrust and power, while a destroyed Calypso was attached to the bottom of the Hab module to provide a cockpit.
Only 2 functional shuttles are present, one is a small escape pod, and the other is the lifeboat-planetary drop pod that can also spawn elsewhere.
Intended for 4 slots. 2 Engineers, 1 Doctor, 1 Pilot.
Weapons are very limited and fitting to the theme.
The Salvage Gantry is supposed to be sketchy and terrifying if used a ship.