Updates Crashed Pod
Finally configures the filters, because players never did that.
Removes clutter, de-crapify a bunch of random stuff.
Drops the SMES charge by 1/5th.
Moves the suits into the spawning room, along with scrubbers and an inflate box, due to the issue of people spawning in after the other room has caught fire.
Surrounds the pod in a destroy turf, due to issues where the pod would require welding to escape.. on a flammable planet.
General shuffling.