Deep Space Fixes
:cl:
bugfix: Deep Space can no longer move to nullspace with people inside of it
/:cl:

"Deep Space" here refers to the lost-in-space mechanic where an empty temporary sector is created. When a player moves to the edge of a temporary sector, a list of all overmap sectors in the same coordinates is created. There is a random chance of moving into any of these colocated sectors, or simply being moved to the other side of the temporary sector.

Previously, all temporary sectors were using a single zlevel. This is noteworthy, because:
- There can be more than one temporary sector instance (imagine two players simultaneously getting lost, but they originate from two different ships (e.g. Torch and Bearcat)
- The overmap code assumes that a zlevel only exists in a single place in the overmap. If this assumption is false, then the zlevel -> sector lookup table `map_sectors` fails to work as expected.
- The temporary sector zlevel was being pulled to the overmap coordinates where a player last entered deep space. This could enable some strange teleportation behavior, where someone enters deep space at overmap coordinates (1,1), then someone else enters deep space at overmap coordinates (25,25). Both players are now in the same zlevel, and are at overmap coordinates (25,25).
- The design of the code heavily implies that each temporary sector instance was supposed to have its own zlevel. It's unclear if the current behavior was always to use a single zlevel across all deep space instances, or if this was a regression introduced at some point.

In this PR, each new temporary sector is allocated a new zlevel. Once a temporary sector has been vacated by all players, it is returned to a pool for later reuse to prevent unbounded creation of new zlevels.

With that explanation behind us, the bug was caused by two things:
- Improper handling of the temporary sector cache, causing their `loc`s to be nulled and never reset.
- The previously explained shared zlevel issue

The effect of the bug was that the temporary sector would effectively be moved to overmap coordinates (0,0) even if players were in it under certain edge cases, effectively stranding everyone in the sector, as it is unlikely there is a ship located at (0,0).