Fixes inconsistency of using shortwaves with ;
There is an inconsistency in which shortwaves will only be talked in
with the shortcut `;` if the user has a headset on their ear that does not
have a working radio connection.
Now even when not having a headset on their ear, shortwaves on a person
will be used with `;`.