Removes the Fire Axe from Bridge Storage
The Torch's bridge has two axes closer together than almost any other place on the ship, with less overall travel time between the two. The placement of the one in bridge storage, right next to two weapons lockers, could make some players believe it is to be used as a defensive weapon. Removing this axe fixes that little problem. And in an emergency where a fireaxe is *actually* necessary, the saferooms should be unbolted and the other fire axe will be available.