Paint sprayer QoL changes
While holding the paint sprayer in an active hand, ctrl-clicking and
shift-clicking on a target are used to pick color or remove color,
respectively (in case of floor tiles, the last applied decal will be
removed).
Includes a new observer pattern implementation for swapping hands in
order to accomplish this behavior.

Other changes:
- The color pick mode and its indicator overlay have been removed (obsolete).
- Ctrl-Click on the paint sprayer will now let you select a color
(instead of swapping into color-pick mode).
- Using the paint sprayer in hand will now immediately let you select a
decal instead of opening a menu.
- Adds a missing decal (full-tile hazard stripes).
- Updated the codex accordingly to reflect the new behavior.