Adds material stall doors
Adds stall doors as a subtype of `door/unpowered/simple`.
Fixes an issue with material doors always being constructed as steel
doors.
Adjusts the hardness bar for furniture recipes to `MATERIAL_FLEXIBLE`
and ups the hardness of plastic to `MATERIAL_FLEXIBLE`.
As a side effect, this will make furniture be buildable from gold,
silver and copper.
Maps the new stall doors in place of curtains in D3 and D1 Head as well as in the medical bathroom.