fix some loadout items not working with JSON saves
For any loadout item that customized paths, if one of the paths had an
\improper name, that path would break.
This is because saving to JSON strips the  \improper part, while
savefiles mainain it.
Fixed by stripping \improper from the internal representation entirely.
I guess this might cause weirdness in the UI if someone tries to do \the
on a loadout entry in the future, but that doesn't happen currently.