Sorts GLOB.player_list by client key
This is done by using ADD_SORTED to add newly logged in players to the
global list. This results in a single O(n) performance hit on login.
This is an infrequent enough event that this should be acceptable, but
may manifest as a transient lag spike with large numbers of players.

This will result in the Lobby player list appearing in alphabetical
order, and may also have the same effect on certain admin tools such as
the debug player list.

As this is done on login, if a mob is transferred to a different client
key, the list will not be updated. As this change primarily targets the
pre-game lobby, before any such changes will be made, I consider this
case to be out of scope.

Apply suggestions from code review

Co-authored-by: Spookerton <spkrtn@pm.me>, Mucker <colehoul@gmail.com>