Ports Vesta's Infantry and Custom Guns 1.0
This merge stills requires somes fixing.

The infantry box has it's Medical and Demolition kits commented due to the glass/termithe and autoinjector/stim bring misscalled considering current Bay.

The Infantry frequency is not working, but the radio and vars are all set.

The Infantry Box sprite is not showing up as intended.

All these notes shall be fixed within the next PR.