Infantry T-Comms setup fix
Note: Mapping issue making it waaay less modular than it should. Instead of making an existing server to handle the Infantry, BoH deleted the Exploration server from T-Comms and added a Infantry Server to handle both Infantry and Exploration.

We want to clean it once we deliver this to live.