General port
I know, thats too much for a PR. Had to raise my errors adoption levels due to the urgency.

It should be faced more as a log of file changes, having in mind that it holds in between these:

-> All BoH's Military Rank logic and var's port
-> All modular mithra, that holds our genemodders, modified unathis and other specimens.
-> All Psykers maths to bring a more Vesta feeling to general Psykers worlds(which is important)
-> A file boh_clothing concentrating BoH's cloak declaration blocks(quite extense).
-> All custom job names.
-> BoHs round rules as avaliable jobs, etc
-> Add's BoH's multiple required defitionitions to run it's own content
-> Ports needed Vesta's macros(file_entry)

And so on. I shall try to documentate it some more.