Patches on voidsuits can now be removed
:cl:
rscadd: Voidsuit patches can now be removed with wirecutters
/:cl:

With this, those annoying patches can finally be removed from voidsuits, allowing a savvy repair person to restore a suit to factory new condition. Previously, the only way a patch could be removed was by attacking the suit's wearer.

I've checked to ensure there aren't any other wirecutter `attackby` interactions that would be broken as a result of this, and everything looks good.

I have not validated this in a local server (yet).