Fix connected pipe coloration issue for pressure tanks
This fixes an edge case that could cause the pipe segment connected to a pressure tank to be the wrong color:
![image](https://user-images.githubusercontent.com/1807509/102004888-76498980-3cda-11eb-9b14-57a3dde70250.png)

Specifically, this happens if a paint sprayer is used to change the tank's color.

Skipping the changelog because the issue is so minor it doesn't need a bugfix tag.