Location Rework
Please do not merge this until it's finished. This is my unfinished location rewrite and expansion.  I'm posting the pr now as a means of getting feedback, and most importantly approval, before committing too much time and effort into this PR.

Contents: Adds a whole bunch of planets to the location system. Probably won't include cultures to match but idk things can change.
Adds a "System" option to locations, so it includes what solar system the planet or location is in.
removes pluto and ceres. I will add them back if needs be but ceres stinky and pluto is a trumpian vision of mexico. Yucky.

Also, this changes up the description of Mars to include the fact it is not wholly terraformed, and changes up Earth to mention how it is a recovered by damaged world. Will probably expand the description. Would be nice if there was a system to split the description shown in character setup and the description shown in Codex, then I could go into more deapth, but c'est la vie.