Fixes absurdly high dislocation chances
It was quickly becoming the worst kept secret that you could dislocate every single one of a persons limbs with a pocket knife in about ten seconds, or both legs in less than one. 

This change makes it so that dislocation is much less likely initially but becomes increasingly more likely to happen as damage is dealt to limbs. 

For reference pre-change you have a whopping 60% chance to dislocate a limb with a pocket knife, post change you'll have an 11% chance, increasing by 11% per hit prior. 

Obviously, higher force weapons still have an exceptionally high chance to dislocate and weapons such as fire axes are almost assuredly still going to displace limbs in one or two hits.

I think this could probably be revisited at a later date but the power of switching to disarm and dislocating limbs is both incredibly high and has little to no method of prevention.