Harpoon & Spear improvements
Both weapons are largely overlooked in favor of more reliable alternatives, this seeks to give them a purpose by allowing them both to stagger and impale characters to walls provided that they're close enough. The damage of being impale to a wall is quite significant, but as the speed of the thrown object remains just barely enough to impale they'll not cause any limbs to randomly explode.