fix observer-related PrivUpdateListenerLoc error
When an observer returns to the lobby or enters a body, it's possible
for them to still be in currently active sound listener event lists,
causing listener_turf to be null if they are picked to hear a sound.

Added a source_turf and listener_turf null skip to prevent this being
a proc crash.