Prebuilt floor_light fix
floor_light/prebuilt would consider itself on (in that it was using power and clicking it says it turns off) but not visually appear to be on at roundstart. This makes floor_light/prebuilt be off and know that it's off and makes a floor_light/prebuilt/on that is definitely and visually on.