Medical uplink rebalancing
The medical uplink had a large number of items that had little to no realistic use cases and lacked many of the important tools required to be able to actually perform the role of medical technician while playing as part of a group.

The goal was to squish the number of available items but increase the usability of the ones that remained and in some cases (sinpockets) replace them with a less fiddly alternative.

First in the list of changes I removed sin pockets and replaced them with a single use injector filled with hyperzine (3u), synaptizine (1u) and inaprovaline (10u).  The injector doesn't facilitate pain immunity alone and instead serves to enhance an antagonists options when it comes to fight or flight, but by no means will it provide them with the ability to ignore incoming fire.

Next up I replaced the stabilization kit with an upgraded slimline version, the slimline can fit into a pocket and takes up much less room in a bag.
The slimline stabilization kit also has an upgraded selection of tools with a focus on allowing an unskilled character to perform emergency first aid in any number of potential situations. It contains
- 1x Advanced trauma kit
- 1x Advanced burn kit
- 1x Splints
- 1x Coagulant*
- 1x autoinjector (Painkiller) 
- 1x emergency low oxygen pouch
- 1x emergency toxin pouch
- 1x emergency radiation pouch
Cost: 16 TC (Unchanged)

*The coagulant injector contains 1u of coagulant and 4u of nanoblood, the coagulant will have a reasonably high chance to heal any internal and external bleeding.

Lastly I removed the health HUD and scanner from the uplink, the individual items cost more than they really should have been valued at and largely bloated the list of objects.

In their place I've added a new "Advanced medical toolkit" which contains the following
- 1x Roller bed
- 1s Box of syringes
- 1x Health HUD
- 1x Health Scanner
- 1x Auto-resuscitator
- 1x Auto-compressor
- 1x Nanoblood IV Bag
- 1x Advanced first-aid kit
- 1x Nitrile Gloves
Cost: 24 TC

Fixes the slimline stabilization kit and some typos


I do not like boxes