Bullet Scatter
Prior to this change bullets could scatter and strike the target in a random location in theory but in order for that to happen the following conditions had to be met

1) You had to miss your attack
2) You had to fail to scatter (Usually a 5% chance to avoid this roll)
3) You had to fail to roll a 70% random probability

This gives us 30% of 5% chance for a bullet that missed to hit a target but not in the chosen location for the shot. I'd imagine the actual odds of this happening are sub 1% once you factor in the hit/miss chance.

With this change in order for a bullet to scatter the following conditions must be met

1) Your attack must hit your target
2) Your attack must then roll a miss on a second roll, the attack will continue to strike the target but now it'll be in a random location.

This means that inaccurate fire is more likely to strike the target in random locations but accurate shots remain accurate. 

Why? Because guns as they currently exist are essentially binary, there's little point wearing any armor on your limbs because hitting those locations is quite unlikely.
This change makes it so that limb armor is far more useful and it's no longer an easy choice to remove it (especially in cases where it added slowdown)
a