Removes the ability to casually dodge bullets
Currently if you move a single tile you get a flat 20% reduction in the chance for a projectile to hit you for a short duration (Even if you run into it!), this chance is reduced if you moved in the same cardinal direction but this really only served to punish people attempting to disengage and flee while it generally advantaged people actively standing to fight. 

Why was this particularly bad?
If a shot had a 100% chance to hit you it'd have an 80% chance to hit if the target had been moving. (a 20% decrease)
If the shot only had a 50/50 to hit (Medium distance, side arm shot) you'd now only have a 30% to be hit, which is a 40% reduction in chance to be hit.

Without going into extremes a slight movement mid combat could make a respectable shot unlikely to strike and make poor shots literally impossible.

I've removed some of the extra calculations from movement, even if moving into a projectile still benefits your ability to dodge it, the chance to evade it is now significantly lower.
The 20% flat evasion has been reduced to 5%.