Lung damage tweaks / bugfixes
I've removed the almost entirely broken code for speaking with lung damage and replaced it with a new system.

Instead of being unable to speak with lung damage and the resulting loss of breath you instead will be forced to whisper. The less breath you have in your lungs the higher the chance of your letters will be replaced by stars until eventually with little to no oxygen in your lungs you'll be basically unable to speak. 

The custom emote character is now ignored when whispering to facilitate this, it's not really the best way of handling it but I couldn't really see any other way to accomplish what I had in mind.