Traitor skill re-mastery
Traitors were gifted a change of one extra master skill at the cost of one less trained skill in #30756, it was later reverted in #30940.

"Bandaid fix for the root issue of master CQC being an obnoxious meme that needs fixing, but far easier than trying to root out the buried code that actually handles this."

Now that master CQC is far more in line with other skills now seems a fair time to give traitors a chance to try out more things.