Adds hangar operations warning beacons, places them in map
Initial Commit

:cl: Textor
rscadd: Warning beacons have been added to the hangar and other relevant locations to inform crew of active docking and undocking operations.
/:cl:
Adds hangar warning beacons. It is designed to be generic enough so that a map maker can get it to work on any map as long as they set up the shuttle correctly and tag the beacon appropriately with the list of shuttles they want to set it off.

Presets are already in place:
- torchhangar for charon, gup, and supply drone
- torchaquilla for the aquilla
-torchpetrov for the petrov

Adds beacon lights to the hangar, the halls outside petrov and aquilla, the EVA area outside the aquilla, the bridge, and the supply office.

Removes beacons from bridge and supply

Added warning beacon to deck 4 hallbridge