Fixes issue where an intellicard slot would block an MFC access panel
:cl: Textor
bugfix: Engineers no longer get confused when trying to do maintenance on a computer containing an intellicard slot.
/:cl:

Fixes #31257

Checks to see if a card is in the slot or not. If it is in the slot, remove the card, if it is not, have the MFC do its normal attackby thing with the screwdriver.

Makes requested changes.