Adds check to see if you replaced eyes to organ attachment end step
:cl: Textor
bugfix: Replacing blinded eyes with undamaged new eyes or undamaged optical sensors now gives you the gift of sight.
/:cl:

Adds a check to the end step of organ attachment in surgery to check if the organ attached was eyes. If it is not in a "broken" state it resets eye_blind value and removes the BLINDED disabled flag. If not in a "bruised" state it will remove the eye_blurry value.

Removes trailing slash