Adds Boron gas to the game
:cl: Textor
rscadd: Adds boron gas. The boron canister can be ordered from supply. Boron is theorized to be good for fusion reactions.
rscadd: Adds gas sensor to RUST chamber.
bugfix: Fixes atmosphere control console in RUST chamber. It now can actually control the injector.
/:cl:

Adds boron to XGM gasses, adjusts the fusion reaction that already referenced boron to use the gas, created a canister to hold the gas, added the canister to the supply list. Boron's stats are based off rounded figures for the real element.

RUST reaction with B and H2 will produce helium, as this is what is produced with actual boron fusion experiments (though the helium lacks electrons in those).

Whitelisted air sensors on the EM field so it won't destabilize the reaction. Added the sensor to the atmos console, fixed the tags so that the atmos console can actually control the injector.

Boron tank is a brown recolor with a B on it.