Restructures the command uniform datums to conform to other departments
:cl: Textor
bugfix: Removes the SEA's ability to be issued a commissioned officer's beret when they are not a commissioned officer.
/:cl:

Restructures the fleet command uniforms to conform to other fleet uniforms, since the "command can only be officers" didn't really take into account that you can be an NCO/SNCO so some equipment might not be appropriate for the rank.

Removed the exception from the uniform vendor code for command.

Tested EC and Fleet command members and received no runtimes or other indications of issues with the uniform vendor after excising that code.

Fixes #31175