Adds Xenobio Cell Areas
:cl: Textor
tweak: Air alarms in xenobiology are now labeled correctly, allowing for engineering to see the cells on the alarm panel and researchers to know which air alarm they are making adjustments to.
maptweak: The power in xenobiology has been recabled to allow for better access to cabling and easier repairs. Each cell also now has an APC.
/:cl:

Adds four new areas for xenobio that distinguishes the different cells. Adds APCs to conform to area requirements. Recables xenobio so that the cables flow better and engineers can restore the area more easily if it's damaged.