Fixes bug with floor attackby code
:cl: Textor
bugfix: Placed tiles should now always be the correct type.
/:cl:

The attackby function for floors was running a for loop through flooring decls. Upon first match, it would set the variable then break the loop. If the first match was a parent object, the child object that should have been selected would not be iterated on, resulting in the wrong tile being placed down.

This removes the break in the loop, allowing the loop to iterate until the end of the decl list. Since this is only called when a plate is attacked by a stack of tiles, I don't think this will have a major performance impact.

Also changes the white monotile turf icon_state to match the appropriate icon_state for the tile.

Changes bugfix solution

Restores the break to the for loop, changes ispath to hard comparison operators.

Removes extra parens

Removes semicolon because this isn't C#