Fixes Traitor Uplink Issue
:cl: Textor
bugfix: Traitor uplinks now quit the program when you use the close button on the popup or in the UI. Closing the window leaves the program running, still.
/:cl:

Adds a program var to /obj/item/device/uplink that references the program set up in /decl/uplink_source/pda/setup_uplink_source so that when the "lock" topic is selected with the close button it was able to use kill_program to shut the program down.

Added logic to the close/resume dialog so that when the player chooses "close" it will toggle the active variable and then permit the code to run the kill_program line that ended the proc.

Fixes #31326