added dmi v5 safety workflow
Byond 515 will see some changes to the .dmi format that
existing tooling like icondiff may be unable to handle
initially. This workflow runs ParadiseSS13/DMI5Checker
against icons/, maps/, and packs/, and will not pass
if a v5 dmi is included in a PR. It can be removed again
once things are up to date.