Give emergency adrenaline autoinjector the intended volume
The emergency adrenaline autoinjector is intended to have 8u of adrenaline, but its volume/capacity was left at the default 5 causing it to only spawn in with 5u. This PR adjusts the volume of the autoinjector so it contains and injects the correct amount, the reason for the increased volume is explained below.

Adrenaline's special "anti-cardiac arrest" effects only apply while someone has more than 4u of adrenaline in their bloodstream. Whoever set this autoinjector up intended them to inject twice that so the effects last longer than a minute, however due to the missing custom volume the adrenaline instead only remains above the critical threshold for seconds. This fix effectively makes adrenaline autoinjectors as effective as they were intended to be.