Ninja PR
adds a new level of depth to the Ninja gamemode in an effort to bring it back into secret. Three new hardsuits for the Ninja mode; a Sol hardsuit, an ICCG hardsuit, and a Corporate hardsuit. Three new hardsuit modules unique to each suit. A remap of the Ninja away site and rebranding to "Operative" as a gamemode.