EMP Damage Scaling
EMP damage now is applied with a random modifier that on average reduces the total damage to humanoid prosthetics and augments. Ion pistols are unlikely to take down an IPC/FBP in a single magazine, but shoot faster to compensate.

EMP damage to limbs has been reduced and flooring stun has been removed to prevent instant immoblization the second someone is hit by an ion or EMP.