Adds NTNRC Notifications and NT Relay Chat to PDA defaults
NTNet Relay Chat now alerts connected channel users to new users, users leaving, and new messages while minimized or while the PDA is in your pocket/bag/ID slot. For sanity's sake, there is a mute button you can use to toggle this on or off.

Adds NTNet Relay Chat to ship PDAs to encourage texting each other as much as emailing each other. Reduces the size of the chat client to 2 GQ because it shouldn't be 8 times the size of an email client.